fix(swarm): swarm network list [BE-13395] (#3569)
