+++ api/http/proxy/factory/docker/transport.go
@@ -22,6 +22,7 @@ import (
 	"github.com/portainer/portainer/api/internal/authorization"
 	"github.com/portainer/portainer/api/logs"
 	"github.com/portainer/portainer/api/slicesx"
+	httprequest "github.com/portainer/portainer/pkg/libhttp/request"
 	"github.com/portainer/portainer/pkg/libhttp/ssrf"
 
 	"github.com/docker/docker/api/types/network"
@@ -30,8 +31,6 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/segmentio/encoding/json"
 )
-
-var apiVersionRe = regexp.MustCompile(`^/v[0-9.]+(/|$)`)
 
 type (
 	// Transport is a custom transport for Docker API reverse proxy. It allows
@@ -138,11 +137,11 @@ func isAdminOnlyRoute(method string, path string) bool {
 func (transport *Transport) ProxyDockerRequest(request *http.Request) (*http.Response, error) {
 	// A percent-encoded path separator lets a request dodge the operation authorization.
 	// Docker API paths never need encoded separators, so reject them outright.
-	if ContainsEncodedSeparator(request.URL.EscapedPath()) {
+	if httprequest.ContainsEncodedSeparator(request.URL.EscapedPath()) {
 		return utils.WriteAccessDeniedResponse()
 	}
 
-	unversionedPath := TrimDockerVersion(request.URL.Path)
+	unversionedPath := httprequest.TrimDockerVersion(request.URL.Path)
 
 	if transport.endpoint.Type == portainer.AgentOnDockerEnvironment || transport.endpoint.Type == portainer.EdgeAgentOnDockerEnvironment {
 		signature, err := transport.signatureService.CreateSignature(portainer.PortainerAgentSignatureMessage)
@@ -979,22 +978,4 @@ func (transport *Transport) fetchEndpointSecuritySettings() (*portainer.Endpoint
 	}
 
 	return &endpoint.SecuritySettings, nil
-}
-
-func ContainsEncodedSeparator(escapedPath string) bool {
-	lower := strings.ToLower(escapedPath)
-	return strings.Contains(lower, "%2f") || strings.Contains(lower, "%5c")
-}
-
-func TrimDockerVersion(urlPath string) string {
-	cleanedPath := path.Clean(urlPath)
-
-	if strings.HasPrefix(cleanedPath, "/v2/") ||
-		strings.HasPrefix(cleanedPath, "v2/") ||
-		strings.HasPrefix(cleanedPath, "/v1/") ||
-		strings.HasPrefix(cleanedPath, "v1/") {
-		return cleanedPath
-	}
-
-	return apiVersionRe.ReplaceAllString(cleanedPath, "/")
 }
